Change packet payload size to uint8_t

+++ src/ks_packet_parser.cpp
@@ -4,7 +4,7 @@ namespace kronos {
 
     static uint32_t s_NextPacketId = 0;
 
-    void EncodePacket(Packet* dest, uint8_t flags, uint16_t command, const uint8_t* payload, uint32_t payloadSize) {
+    void EncodePacket(Packet* dest, uint8_t flags, uint16_t command, const uint8_t* payload, uint8_t payloadSize) {
         *dest = {
             .Header = {
                 .Magic = KSP_MAGIC,
